requirements.txt
numpy~=1.26.4
gymnasium~=1.1.1
stable_baselines3~=2.7.1
torch~=2.3.1+cu121
ray~=2.52.1
wandb~=0.17.5
tqdm~=4.66.4
hydra-core~=1.3.2
omegaconf~=2.3.0
